Added run.
Added method to run the fire.  But to get the callback working we had to
wrap the class burn call with a locally defined method.  Also we needed
to call separate classes of the burn method.  It would appear a better
way would be to have a single burn method & call back to handle all LED
functions.