Call time.run from loop().
Buy using externs we can access the time object methods from inside the
fire objects with out having to get a pointer to the time methods.  This
freed up the instantiation of the time object to be done in the sketch
and so we can call the time.run method from the sketch directly.